removed warning
git-svn-id: http://svn.seqan.de/seqan/trunk@8976 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/apps/razers3/razers_matepairs_parallel.h
@@ -1452,10 +1452,10 @@ int _mapMatePairReadsParallel(
         options.countFiltration += threadLocalStorages[i].options.countFiltration;
         options.countVerification += threadLocalStorages[i].options.countVerification;
     }
-	
+
 	// restore original orientation (R-reads are infixes of ConcatDirect StringSet)
 	#pragma omp parallel for schedule(static, 1)
-	for (int i = 0; i < length(threadLocalStorages); ++i)
+	for (int i = 0; i < (int)length(threadLocalStorages); ++i)
 		reverseComplement(threadLocalStorages[i].readSetR);
 
 	options.timeMapReads = SEQAN_PROTIMEDIFF(findTime);
